+++ base/MCRun.py
@@ -29,6 +29,7 @@ class MCRun:
     base.VERSION.added("1.7.0", "Type hints to `base.MCRun` ")
     base.VERSION.changed("1.8.0", "Replaced Legacy format strings by f-strings in `base.MCRun` ")
     base.VERSION.added("1.10.0", "XML-tag for element names in `base.MCRun` configurations")
+    base.VERSION.changed("1.13.0", "Allowed more spellings for enabling/disabling components in `base.MCRun` ")
 
     def __init__(self, xml_file: str, **keywords) -> None:
         config = xml.etree.ElementTree.parse(xml_file)
@@ -53,7 +54,7 @@ class MCRun:
                     f"Component {componentConfig.tag} already present in data store",
                     "Configuration and parameterization ignored"
                 )
-            elif ("enabled" not in componentConfig.attrib or componentConfig.attrib["enabled"] == "true") and \
+            elif ("enabled" not in componentConfig.attrib or componentConfig.attrib["enabled"].lower() == "true") and \
                     ("enabled_expression" not in componentConfig.attrib or
                      eval(componentConfig.attrib["enabled_expression"])):
                 component_module = importlib.import_module(componentConfig.attrib["module"])