+++ base/documentation.py
@@ -44,6 +44,7 @@ base.VERSION.fixed("1.12.2", "Typos in `base.documentation` ")
 base.VERSION.added("1.12.2", "Step to merge request instructions in `base.documentation` ")
 base.VERSION.changed("1.12.5", "Enhanced contribution documentation in `base.documentation` ")
 base.VERSION.changed("1.12.6", "Refactored documentation of class members into own function in `base.documentation` ")
+base.VERSION.fixed("1.13.0", "Spelling error in `base.documentation` ")
 
 
 def write_changelog(name: str, version_history: base.VersionCollection, file_path: str) -> None:
@@ -376,7 +377,7 @@ please adhere to the following steps:
 If you are requesting a merge containing component code, please make sure that the following applies:
 - [ ] The commit that is requested to branch is based on the most recent commit on the master branch.
 - [ ] The repository can be cloned from GitLab.
-- [ ] The component runs successfully without any errors using the most recent model version..
+- [ ] The component runs successfully without any errors using the most recent model version.
 - [ ] You haven't reverted any changes made by other contributors unless there is a good reason to do so.
 - [ ] You haven't introduced inputs to the component that are not needed for calculations.
 
@@ -384,7 +385,7 @@ If you are requesting a merge containing component code, please make sure that t
 If you are requesting a merge relating to a model variant, please make sure that the following applies:
 - [ ] The commit that is requested to branch is based on the most recent commit on the master branch.
 - [ ] The entire model, including all submodules, can be cloned from GitLab.
-- [ ] The model runs successfully without any errors using the most recent model version..
+- [ ] The model runs successfully without any errors using the most recent model version.
 - [ ] You haven't reverted any changes made by other contributors unless there is a good reason to do so.
 
 ## Scenario